same-chunk attention bf16-mixed 좌표 해상도 손실 수정
``bf16-mixed`` 학습 경로에서 ``chunk_tokens.dtype`` 은 ``bfloat16`` 이지만,
``forward`` 가 ``interaction_pos`` / ``interaction_head`` 를 같은 dtype 으로
캐스팅해 graph 구성 좌표가 bf16 으로 내려갔습니다. Waymo 좌표는 수천~수만 m
까지 커질 수 있어 bf16 캐스팅 시 좌표 해상도가 m 단위 이상으로 거칠어져
60m radius graph 경계 근처 엣지가 빠지거나 잘못 생겼습니다. 학습 시점마다
graph topology 가 흔들리며 WOSAC RMM 에 직접 악영향을 줄 수 있는 수치적
하자입니다.

``forward`` 가 ``interaction_pos`` / ``interaction_head`` 를 항상 ``float32`` 로
끌어올려 ``_build_same_chunk_edges`` 에 넘기도록 수정했습니다. 이후 Fourier
embedding 안의 ``nn.Linear`` 가 autocast 로 결과를 attention dtype 에 맞게
캐스팅하므로 attention 단 dtype 일관성은 그대로 유지됩니다.

bf16 토큰과 호출자가 잘못 bf16 으로 캐스팅한 좌표 두 경우 모두 graph 경로
좌표가 fp32 로 유지되는지 확인하는 회귀 테스트 두 개를 추가했습니다. 수정
전에는 두 테스트 모두 실패하고, 수정 후 모두 통과합니다. 4xH100 ``bf16-mixed``
종단 smoke 학습도 정상 완료(NaN/예외 없음, loss·RMM·ADE 모두 호전)했습니다.

+++ src/smart/modules/flow_local_decoder.py
@@ -732,8 +732,19 @@ class SameChunkAgentAttentionBlock(nn.Module):
             num_anchor=num_anchor,
             device=chunk_tokens.device,
         )
-        interaction_pos = interaction_pos.to(device=chunk_tokens.device, dtype=chunk_tokens.dtype)
-        interaction_head = interaction_head.to(device=chunk_tokens.device, dtype=chunk_tokens.dtype)
+        # ``bf16-mixed`` 학습에서 ``chunk_tokens.dtype`` 은 ``bfloat16`` 이지만,
+        # Waymo 좌표는 수천~수만 m 까지 커질 수 있어 bf16 으로 내리면 위치
+        # 해상도가 m 단위 이상으로 거칠어집니다. 그러면 60m radius graph 의
+        # 엣지가 빠지거나 잘못 생겨 학습 시 graph topology 가 흔들립니다.
+        # 따라서 graph 구성과 edge feature 계산에 쓰이는 좌표·heading 은
+        # 항상 float32 로 유지하고, 이후 Fourier embedding 의 nn.Linear 가
+        # autocast 로 attention dtype 에 맞게 결과를 캐스팅하도록 둡니다.
+        interaction_pos = interaction_pos.to(
+            device=chunk_tokens.device, dtype=torch.float32
+        )
+        interaction_head = interaction_head.to(
+            device=chunk_tokens.device, dtype=torch.float32
+        )
         edge_index, edge_attr = self._build_same_chunk_edges(
             interaction_pos=interaction_pos,
             interaction_head=interaction_head,
